Major Refactor

+++ controller/comments.go
@@ -1,48 +1,55 @@
 package controller
 
 import (
-	"encoding/json"
 	"errors"
 	"net/http"
+	"strconv"
 
-	"github.com/gorilla/mux"
+	"github.com/labstack/echo/v4"
 	"gorm.io/gorm"
 	"readyworker.com/backend/database"
 	"readyworker.com/backend/model"
 )
 
-func PutComment(w http.ResponseWriter, r *http.Request) {
-	var comment model.Comment
-	err := json.NewDecoder(r.Body).Decode(&comment)
-
-	if err != nil {
-		http.Error(w, "Error decoding request", http.StatusBadRequest)
+func CreateComment(c echo.Context) (err error) {
+	comment := new(model.Comment)
+	if err = c.Bind(comment); err != nil {
 		return
 	}
+
+	comment.AuthorID = GetIDFromToken(c)
+	role := GetRoleFromToken(c)
+
+	// Validation
+	if comment.Content == "" || comment.WorkerID == 0 {
+		return &echo.HTTPError{Code: http.StatusBadRequest, Message: "Invalid content or id fields"}
+	}
+	if role != "hirer" {
+		return &echo.HTTPError{Code: http.StatusUnauthorized, Message: "Unauthorized comment"}
+	}
+
+	// TODO: Check if the hirer hired the worker
 
 	db := database.Connect()
 	defer database.Disconnect(db)
-	// TODO: Authenticate
+
 	var author model.User
 	result := db.Where("ID = ?", comment.AuthorID).First(&author)
 	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		http.Error(w, "Author does not exists", http.StatusBadRequest)
-		//http.Redirect(w, r, "/login?retry=1", http.StatusPermanentRedirect)
-		return
+		return &echo.HTTPError{Code: http.StatusBadRequest, Message: "Author does not exists"}
 	}
 	result = db.Where("ID = ?", comment.WorkerID).First(&author)
 	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		http.Error(w, "Worker does not exists", http.StatusBadRequest)
-		//http.Redirect(w, r, "/login?retry=1", http.StatusPermanentRedirect)
-		return
+		return &echo.HTTPError{Code: http.StatusBadRequest, Message: "Worker does not exists"}
 	}
 
 	db.Create(&comment)
 
+	return c.JSON(http.StatusCreated, comment)
+
 }
-func DeleteComment(w http.ResponseWriter, r *http.Request) {
-	// TODO: Authenticate
-	commentID := mux.Vars(r)["id"]
+func DeleteComment(c echo.Context) error {
+	commentID, _ := strconv.Atoi(c.Param("id"))
 
 	db := database.Connect()
 	defer database.Disconnect(db)
@@ -50,20 +57,22 @@ func DeleteComment(w http.ResponseWriter, r *http.Request) {
 	var comment model.Comment
 	result := db.Where("ID = ?", commentID).First(&comment)
 	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		http.Error(w, "Comment Not Found", http.StatusNotFound)
-		//http.Redirect(w, r, "/login?retry=1", http.StatusPermanentRedirect)
-		return
+		return &echo.HTTPError{Code: http.StatusNotFound, Message: "Comment Not Found"}
 	}
+
+	userID := GetIDFromToken(c)
+	role := GetRoleFromToken(c)
+
+	if comment.AuthorID != userID && role != "admin" {
+		return &echo.HTTPError{Code: http.StatusUnauthorized, Message: "Unauthorized delete"}
+	}
+
 	db.Delete(&comment)
 
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"status": "success",
-	})
+	return c.NoContent(http.StatusNoContent)
 }
-func GetComment(w http.ResponseWriter, r *http.Request) {
-	// TODO: Authenticate
-	commentID := mux.Vars(r)["id"]
+func GetComment(c echo.Context) error {
+	commentID, _ := strconv.Atoi(c.Param("id"))
 
 	db := database.Connect()
 	defer database.Disconnect(db)
@@ -71,12 +80,9 @@ func GetComment(w http.ResponseWriter, r *http.Request) {
 	var comment model.Comment
 	result := db.Where("ID = ?", commentID).First(&comment)
 	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		http.Error(w, "Comment Not Found", http.StatusNotFound)
-		//http.Redirect(w, r, "/login?retry=1", http.StatusPermanentRedirect)
-		return
+		return &echo.HTTPError{Code: http.StatusNotFound, Message: "Comment Not Found"}
 	}
 
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(&comment)
+	return c.JSON(http.StatusOK, comment)
 
 }
